WeeklyStorage: allow tracking of a single highest value within the week
Example use case is if we have a feature that has a concept of a session and we want to track the largest number of interactions in that session within the previous week (or p3a tracking period). REplaceTodaysValueIfGreater combined with GetHighestValueInWeek allows us to retrieve the session with the highest number of interactions, whilst also expiring sessions more than 7 days old.